fix: there were cases of accessing null fields and the package was refactored
