+++ pkg/api/personal_token/client.go
@@ -1,0 +1,29 @@
+package personal_token
+
+import (
+	"net/http"
+	"time"
+
+	"github.com/aziontech/azion-cli/pkg/cmd/version"
+	sdk "github.com/aziontech/azionapi-go-sdk/personal_tokens"
+)
+
+type Client struct {
+	apiClient *sdk.APIClient
+}
+
+func NewClient(c *http.Client, url string, token string) *Client {
+	conf := sdk.NewConfiguration()
+	conf.HTTPClient = c
+	conf.AddDefaultHeader("Authorization", "token "+token)
+	conf.AddDefaultHeader("Accept", "application/json;version=3")
+	conf.UserAgent = "Azion_CLI/" + version.BinVersion
+	conf.Servers = sdk.ServerConfigurations{
+		{URL: url},
+	}
+	conf.HTTPClient.Timeout = 30 * time.Second
+
+	return &Client{
+		apiClient: sdk.NewAPIClient(conf),
+	}
+}
