+++ pkg/api/personal_token/create.go
@@ -2,36 +2,13 @@ package personal_token
 
 import (
 	"context"
-	"net/http"
 	"time"
 
-	"github.com/aziontech/azion-cli/pkg/cmd/version"
 	"github.com/aziontech/azion-cli/pkg/logger"
 	"github.com/aziontech/azion-cli/utils"
-	"go.uber.org/zap"
-
 	sdk "github.com/aziontech/azionapi-go-sdk/personal_tokens"
+	"go.uber.org/zap"
 )
-
-type Client struct {
-	apiClient *sdk.APIClient
-}
-
-func NewClient(c *http.Client, url string, token string) *Client {
-	conf := sdk.NewConfiguration()
-	conf.HTTPClient = c
-	conf.AddDefaultHeader("Authorization", "token "+token)
-	conf.AddDefaultHeader("Accept", "application/json;version=3")
-	conf.UserAgent = "Azion_CLI/" + version.BinVersion
-	conf.Servers = sdk.ServerConfigurations{
-		{URL: url},
-	}
-	conf.HTTPClient.Timeout = 30 * time.Second
-
-	return &Client{
-		apiClient: sdk.NewAPIClient(conf),
-	}
-}
 
 type Response interface {
 	GetUuid() string
@@ -55,10 +32,6 @@ func (c *Client) Create(ctx context.Context, req *Request) (Response, error) {
 
 	if err != nil {
 		logger.Error("Error while creating a personal token", zap.Error(err))
-		logger.Debug("Status Code", zap.Any("http", httpResp.StatusCode))
-		logger.Debug("Headers", zap.Any("http", httpResp.Header))
-		logger.Debug("Response body", zap.Any("http", httpResp.Body))
-
 		return nil, utils.ErrorPerStatusCode(httpResp, err)
 	}
